+++ partiture-example-contract/src/main/kotlin/mypackage/cordapp/contract/AccountYoContractAndState.kt
@@ -12,15 +12,20 @@
  *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *     Lesser General Public License for more details.
  *
- *     You should have received a copy of the GNU Lesser General Public
+ *     AccountYou should have received a copy of the GNU Lesser General Public
  *     License along with this library; if not, write to the Free Software
  *     Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
  *     USA
  */
 package mypackage.cordapp.contract
 
-import net.corda.core.contracts.*
-import net.corda.core.identity.Party
+import com.github.manosbatsis.vaultaire.dto.AccountParty
+import net.corda.core.contracts.BelongsToContract
+import net.corda.core.contracts.Contract
+import net.corda.core.contracts.ContractState
+import net.corda.core.contracts.TypeOnlyCommandData
+import net.corda.core.contracts.requireSingleCommand
+import net.corda.core.contracts.requireThat
 import net.corda.core.schemas.MappedSchema
 import net.corda.core.schemas.PersistentState
 import net.corda.core.schemas.QueryableState
@@ -31,10 +36,10 @@ import javax.persistence.Table
 
 
 // Contract and state.
-val YO_CONTRACT_PACKAGE = YoContract::class.java.`package`.name
-val YO_CONTRACT_ID = YoContract::class.java.canonicalName
+val ACCOUNT_YO_CONTRACT_PACKAGE = AccountYoContract::class.java.`package`.name
+val ACCOUNT_YO_CONTRACT_ID = AccountYoContract::class.java.canonicalName
 
-class YoContract : Contract {
+class AccountYoContract : Contract {
 
     // Command.
     class Send : TypeOnlyCommandData()
@@ -42,32 +47,32 @@ class YoContract : Contract {
     // Contract code.
     override fun verify(tx: LedgerTransaction) = requireThat {
         val command = tx.commands.requireSingleCommand<Send>()
-        "There can be no inputs when Yo'ing other parties." using (tx.inputs.isEmpty())
-        "There must be one output: The Yo!" using (tx.outputs.size == 1)
-        val yo = tx.outputsOfType<YoState>().single()
-        "No sending Yo's to yourself!" using (yo.recepient != yo.sender)
-        "The Yo! must be signed by the sender." using (command.signers.contains(yo.sender.owningKey))
-        //"The Yo! must be signed by the recipient." using (command.signers.contains(yo.recepient.owningKey))
+        "There can be no inputs when AccountYo'ing other parties." using (tx.inputs.isEmpty())
+        //"There must be one output: The AccountYo!" using (tx.outputs.size == 1)
+        val yo = tx.outputsOfType<AccountYoState>().first()
+        "No sending AccountYo's to yourself!" using (yo.recepient != yo.sender)
+        "The AccountYo! must be signed by the sender." using (command.signers.contains(yo.sender.party.owningKey))
+        //"The AccountYo! must be signed by the recipient." using (command.signers.contains(yo.recepient.owningKey))
     }
 
     // State.
-    @BelongsToContract(YoContract::class)
-    data class YoState(val sender: Party,
-                       val recepient: Party,
-                       val yo: String = "Yo!") : ContractState, QueryableState {
-        override val participants get() = listOf(sender, recepient)
+    @BelongsToContract(AccountYoContract::class)
+    data class AccountYoState(val sender: AccountParty,
+                              val recepient: AccountParty,
+                              val yo: String = "AccountYo!") : ContractState, QueryableState {
+        override val participants get() = listOf(sender.party, recepient.party)
         //override fun toString() = "${sender.name}: $yo"
-        override fun supportedSchemas() = listOf(YoSchemaV1)
+        override fun supportedSchemas() = listOf(AccountYoSchemaV1)
 
-        override fun generateMappedObject(schema: MappedSchema) = YoSchemaV1.PersistentYoState(
+        override fun generateMappedObject(schema: MappedSchema) = AccountYoSchemaV1.PersistentAccountYoState(
                 sender.name.toString(), recepient.name.toString(), yo)
 
-        object YoSchema
+        object AccountYoSchema
 
-        object YoSchemaV1 : MappedSchema(YoSchema.javaClass, 1, listOf(PersistentYoState::class.java)) {
+        object AccountYoSchemaV1 : MappedSchema(AccountYoSchema.javaClass, 1, listOf(PersistentAccountYoState::class.java)) {
             @Entity
             @Table(name = "yos")
-            class PersistentYoState(
+            class PersistentAccountYoState(
                     @Column(name = "sender")
                     var origin: String = "",
                     @Column(name = "recepient")